Move ui to frontend-core
